# contracts/openapi/v1.yaml
openapi: 3.0.3
info:
  title: MYJUNGLE Backend API
  version: 1.3.0
  description: |
    Contract for backoffice and MCP integrations.
    This document tracks the current `backend-api` router surface, including:
    session auth, platform management, SSH/API key management, project analysis,
    commit history, SSE streams, and both embedding and LLM provider settings.

    ## Error Status Codes
    - `400` - invalid JSON body or malformed query/path parameters
    - `401` - authentication required or invalid credentials
    - `403` - insufficient permissions
    - `404` - resource not found
    - `409` - conflict
    - `413` - request body too large
    - `422` - validation error
    - `500` - internal server error
    - `502` - upstream dependency unavailable
servers:
  - url: http://localhost:8080
    description: Local
security:
  - BearerAuth: []
  - CookieAuth: []
...
paths:
...
  /v1/auth/login:
    post:
      tags: [Auth]
      operationId: login
      security: []
      description: |
        Phase 1 uses username-only authentication on trusted local networks.
        Creates a server-side session and returns the raw token plus a Set-Cookie header.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          headers:
            Set-Cookie:
              description: Session cookie (`mj_session`)
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "422":
          $ref: "#/components/responses/ErrorResponse"
...
  /v1/ssh-keys:
    get:
      tags: [SSH Keys]
      operationId: listSSHKeys
      responses:
        "200":
          description: SSH keys owned by the current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SSHKeyListResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
    post:
      tags: [SSH Keys]
      operationId: createSSHKey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSSHKeyRequest"
      responses:
        "201":
          description: SSH key created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SSHKey"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "422":
          $ref: "#/components/responses/ErrorResponse"
  /v1/ssh-keys/{ssh_key_id}:
    parameters:
      - $ref: "#/components/parameters/SSHKeyID"
    get:
      tags: [SSH Keys]
      operationId: getSSHKey
      responses:
        "200":
          description: SSH key detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SSHKey"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "422":
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
      description: Session token or API key via the `Authorization` header
    CookieAuth:
      type: apiKey
      in: cookie
      name: mj_session
      description: Session cookie set by `POST /v1/auth/login`
  parameters:
...
    SSHKeyID:
      in: path
      name: ssh_key_id
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    ErrorResponse:
      description: Structured error payload
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
          additionalProperties: true
...
    LoginRequest:
      type: object
      required: [username]
      properties:
        username:
          type: string
    User:
      type: object
      required: [id, username, email, is_active, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        display_name:
          type: string
        avatar_url:
          type: string
          format: uri
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    LoginResponse:
      type: object
      required: [token, expires_at, user]
      properties:
        token:
          type: string
        expires_at:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
...
    SSHKeySummary:
      type: object
      required: [id, name, fingerprint, public_key, key_type, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        fingerprint:
          type: string
        public_key:
          type: string
        key_type:
          type: string
        created_at:
          type: string
          format: date-time
    SSHKey:
      allOf:
        - $ref: "#/components/schemas/SSHKeySummary"
        - type: object
          properties:
            is_active:
              type: boolean
            created_by:
              type: string
              format: uuid
            rotated_at:
              type: string
              format: date-time
              nullable: true
    SSHKeyListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SSHKey"
...
    CreateSSHKeyRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 100
        private_key:
          type: string
          description: PEM-encoded private key to import. Omit to generate a new key pair.
...